+++ tests/transport/graphQL/test_graphQL.py
@@ -4,8 +4,36 @@ from typing import Any
 
 import pytest
 from fastapi.testclient import TestClient
+from pytest_mock import MockerFixture
+from tests.assertable_controller import (
+    AssertableController,
+    TestHandler,
+    TestSender,
+    TestUpdater,
+)
 
+from fastcs.attributes import AttrR, AttrRW, AttrW
+from fastcs.datatypes import Bool, Float, Int, String
 from fastcs.transport.graphQL.adapter import GraphQLTransport
+
+
+class RestAssertableController(AssertableController):
+    read_int = AttrR(Int(), handler=TestUpdater())
+    read_write_int = AttrRW(Int(), handler=TestHandler())
+    read_write_float = AttrRW(Float())
+    read_bool = AttrR(Bool())
+    write_bool = AttrW(Bool(), handler=TestSender())
+    read_string = AttrRW(String())
+    big_enum = AttrR(
+        Int(
+            allowed_values=list(range(17)),
+        ),
+    )
+
+
+@pytest.fixture(scope="class")
+def assertable_controller(class_mocker: MockerFixture):
+    return RestAssertableController(class_mocker)
 
 
 def nest_query(path: list[str]) -> str:
@@ -44,10 +72,12 @@ def nest_responce(path: list[str], value: Any) -> dict:
 class TestGraphQLServer:
     @pytest.fixture(scope="class")
     def client(self, assertable_controller):
-        app = GraphQLTransport(assertable_controller)._server._app
+        app = GraphQLTransport(
+            assertable_controller,
+        )._server._app
         return TestClient(app)
 
-    def test_read_int(self, assertable_controller, client):
+    def test_read_int(self, client, assertable_controller):
         expect = 0
         path = ["readInt"]
         query = f"query {{ {nest_query(path)} }}"
@@ -56,7 +86,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, expect)
 
-    def test_read_write_int(self, assertable_controller, client):
+    def test_read_write_int(self, client, assertable_controller):
         expect = 0
         path = ["readWriteInt"]
         query = f"query {{ {nest_query(path)} }}"
@@ -72,7 +102,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, new)
 
-    def test_read_write_float(self, assertable_controller, client):
+    def test_read_write_float(self, client, assertable_controller):
         expect = 0
         path = ["readWriteFloat"]
         query = f"query {{ {nest_query(path)} }}"
@@ -88,7 +118,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, new)
 
-    def test_read_bool(self, assertable_controller, client):
+    def test_read_bool(self, client, assertable_controller):
         expect = False
         path = ["readBool"]
         query = f"query {{ {nest_query(path)} }}"
@@ -97,7 +127,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, expect)
 
-    def test_write_bool(self, assertable_controller, client):
+    def test_write_bool(self, client, assertable_controller):
         value = True
         path = ["writeBool"]
         mutation = f"mutation {{ {nest_mutation(path, value)} }}"
@@ -106,23 +136,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, value)
 
-    def test_string_enum(self, assertable_controller, client):
-        expect = ""
-        path = ["stringEnum"]
-        query = f"query {{ {nest_query(path)} }}"
-        with assertable_controller.assert_read_here(["string_enum"]):
-            response = client.post("/graphql", json={"query": query})
-        assert response.status_code == 200
-        assert response.json()["data"] == nest_responce(path, expect)
-
-        new = "new"
-        mutation = f"mutation {{ {nest_mutation(path, new)} }}"
-        with assertable_controller.assert_write_here(["string_enum"]):
-            response = client.post("/graphql", json={"query": mutation})
-        assert response.status_code == 200
-        assert response.json()["data"] == nest_responce(path, new)
-
-    def test_big_enum(self, assertable_controller, client):
+    def test_big_enum(self, client, assertable_controller):
         expect = 0
         path = ["bigEnum"]
         query = f"query {{ {nest_query(path)} }}"
@@ -131,7 +145,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, expect)
 
-    def test_go(self, assertable_controller, client):
+    def test_go(self, client, assertable_controller):
         path = ["go"]
         mutation = f"mutation {{ {nest_query(path)} }}"
         with assertable_controller.assert_execute_here(path):
@@ -139,7 +153,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == {path[-1]: True}
 
-    def test_read_child1(self, assertable_controller, client):
+    def test_read_child1(self, client, assertable_controller):
         expect = 0
         path = ["SubController01", "readInt"]
         query = f"query {{ {nest_query(path)} }}"
@@ -148,7 +162,7 @@ class TestGraphQLServer:
         assert response.status_code == 200
         assert response.json()["data"] == nest_responce(path, expect)
 
-    def test_read_child2(self, assertable_controller, client):
+    def test_read_child2(self, client, assertable_controller):
         expect = 0
         path = ["SubController02", "readInt"]
         query = f"query {{ {nest_query(path)} }}"
