+++ tests/transport/tango/test_dsr.py
@@ -1,10 +1,22 @@
 import asyncio
 from unittest import mock
 
+import enum
+
+import numpy as np
 import pytest
+from pytest_mock import MockerFixture
 from tango import DevState
 from tango.test_context import DeviceTestContext
+from tests.assertable_controller import (
+    AssertableController,
+    TestHandler,
+    TestSender,
+    TestUpdater,
+)
 
+from fastcs.attributes import AttrR, AttrRW, AttrW
+from fastcs.datatypes import Bool, Enum, Float, Int, String, WaveForm
 from fastcs.transport.tango.adapter import TangoTransport
